requirements.txt
anyio==3.6.2
base58==2.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptoconditions==0.8.1
cryptography==3.4.7
Deprecated==1.2.13
ECPy==1.2.5
h11==0.14.0
httpcore==0.16.3
httpx==0.23.2
idna==3.4
pyasn1==0.4.8
pycparser==2.21
PyNaCl==1.4.0
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
types-Deprecated==1.2.9
typing_extensions==4.4.0
urllib3==1.26.13
websockets==10.4
wrapt==1.14.1
xrpl-py==1.8.0b0